Move SemanticVersion to shared location
