+++ packages/devtools_app/lib/src/framework/release_notes/release_notes.dart
@@ -5,6 +5,7 @@
 import 'dart:async';
 import 'dart:math';
 
+import 'package:devtools_shared/devtools_shared.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
@@ -17,7 +18,6 @@ import '../../config_specific/server/server.dart' as server;
 import '../../primitives/auto_dispose_mixin.dart';
 import '../../shared/common_widgets.dart';
 import '../../shared/theme.dart';
-import '../../shared/version.dart';
 
 const debugTestReleaseNotes = false;
