+++ packages/devtools_app/test/shared/visible_screens_test.dart
@@ -23,6 +23,7 @@ import 'package:devtools_app/src/shared/globals.dart';
 import 'package:devtools_app/src/shared/preferences.dart';
 import 'package:devtools_app/src/shared/screen.dart';
 import 'package:devtools_app/src/shared/version.dart';
+import 'package:devtools_shared/devtools_shared.dart';
 import 'package:devtools_test/devtools_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
